releases/tars-comm-dev/registry/requirements.txt
fastapi
python-multipart
requests
python-dotenv
supabase
uvicorn
Jinja2